# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is URL shortener API",
        "title": "Shortener API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/RouSage/shortener/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
        "/v1/admin/urls": {
            "get": {
                "description": "Retrieves a paginated list of all URLs created by users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all URLs",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Get custom URLs only",
                        "name": "isCustom",
                        "in": "query"
                    },
                    {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string",
                        "description": "Get URLs created by a specific user",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "maximum": 10000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of URLs",
                        "schema": {
                            "$ref": "#/definitions/server.PaginatedURLs"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/v1/admin/urls/user/{userId}": {
            "delete": {
                "description": "Delete all URLs created by a user. Also removes them from cache.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete URLs created by a user",
                "parameters": [
                    {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string",
                        "description": "ID of the user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - URLs successfully deleted"
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/v1/admin/urls/{code}": {
            "delete": {
                "description": "Deletes a URL. Also removes it from cache.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete URL",
                "parameters": [
                    {
                        "maxLength": 16,
                        "type": "string",
                        "description": "Short code of the URL",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - URL successfully deleted"
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Short URL not found",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "appvalidator.ValidationError": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "repository.Url": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isCustom": {
                    "type": "boolean"
                },
                "longUrl": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "server.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "server.HTTPValidationError": {
            "type": "object",
            "properties": {
                "errors": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/appvalidator.ValidationError"
                        }
                    ],
                    "example": {
                        "{\"field\"": "\"field error message\"}"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "server.PaginatedURLs": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.Url"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/server.Pagination"
                }
            }
        },
...
        "server.Pagination": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "type": "boolean"
                },
                "hasPrevious": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}